% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSpider.R
\name{multiSpider}
\alias{multiSpider}
\title{multispider plot}
\usage{
multiSpider(x1 = NULL, x2 = NULL, colRec = NULL, alphaRec = 0.5,
  colRecBorder = "#ff6666", titles = NULL, cexSteps = c(0.01, 1.4),
  cexProcent = c(0.2, 0.8), parValues = list(pty = "s"),
  circleLines = c(5, 5), mar = list(c(0, 2, 1, 2) + 0.7, c(0, 0, 1, 0) +
  0.2), cexPoints = 1, sizeMiddle = 0.5, twist = 90, multiCoords = NULL,
  ...)
}
\arguments{
\item{x1}{data}

\item{x2}{data to plot on top of x1}

\item{colRec}{color of x1 rectangular}

\item{alphaRec}{alpha of the rectangular}

\item{colRecBorder}{color of rectangular's border}

\item{titles}{for each plot}

\item{cexSteps}{size of radial labels}

\item{cexProcent}{size of procent texts}

\item{parValues}{par(...) values to be called for each plot}

\item{circleLines}{default = 5}

\item{mar}{list of 2xmargin vectors, first one for the inner plot}

\item{cexPoints}{size of the points}

\item{sizeMiddle}{proportion of middle plot to the outer plots}

\item{twist}{twist the outer plots}

\item{...}{arguments passend down to spider()}
}
\description{
multispider plot
}
