% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spider.R
\name{spider}
\alias{spider}
\title{Spider chart function}
\usage{
spider(x1 = NULL, x2 = NULL, colRec = "#ff9999", alphaRec = 0.5,
  colRecBorder = NULL, titles = NULL, stepsText = NULL, singlePanel = F,
  rectangular = F, colSpider = "#e6e6e6", alphaSpider = 0.5,
  colBorder = c("#cccccc", "#cccccc", "#666798", "#cccccc", "#cccccc"),
  maxValues = NULL, minValues = NULL, rad = 5, cexSteps = 1.4,
  cexProcent = 0.8, parValues = list(pty = "s"), cexPoints = 1)
}
\arguments{
\item{x1}{measurement data}

\item{x2}{color for spider}

\item{maxValues}{alpha for colSpider}

\item{minValues}{colors for spider borders}
}
\description{
Spider chart function
}
