% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSpider.R
\name{multiSpider}
\alias{multiSpider}
\title{multispider plot}
\usage{
multiSpider(x1 = NULL, x2 = NULL, colRec = NULL, alphaRec = 0.5,
  colRecBorder = "#ff6666", titles = NULL, cexSteps = c(0.01, 1.4),
  cexProcent = c(0.2, 0.8), parValues = list(pty = "s"), mar = list(c(0,
  2, 1, 2) + 0.7, c(0, 0, 1, 0) + 0.2), cexPoints = 1, sizeMiddle = 0.5,
  twist = 90, ...)
}
\arguments{
\item{twist}{}
}
\description{
multispider plot
}
