% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spider.R
\name{spider}
\alias{spider}
\title{Spider chart function}
\usage{
spider(x1 = NULL, x2 = NULL, colRec = "#ff9999", alphaRec = 0.5,
  colRecBorder = NULL, titles = NULL, stepsText = NULL, singlePanel = F,
  rectangular = F, colSpider = "#e6e6e6", alphaSpider = 0.5,
  colBorder = c("#cccccc", "#cccccc", "#666798", "#cccccc", "#cccccc"),
  maxValues = NULL, minValues = NULL, rad = 5, cexSteps = 1.4,
  cexProcent = 0.8, parValues = list(pty = "s"), cexPoints = 1)
}
\arguments{
\item{x1}{data}

\item{x2}{data to plot on top of x1}

\item{colRec}{color of x1 rectangular}

\item{alphaRec}{alpha of the rectangular}

\item{colRecBorder}{color of rectangular's border}

\item{titles}{title for the spider plot}

\item{stepsText}{labels for the radar lines}

\item{singlePanel}{plot everything in one plot}

\item{rectangular}{rectangular spider chart}

\item{colSpider}{color of the spider chart}

\item{alphaSpider}{alpha of the spider chart}

\item{colBorder}{color for the five rings of the chart}

\item{maxValues}{which x1/x2 maximal can get}

\item{minValues}{which x1/x2 minimal can get}

\item{rad}{radius for the charts, plot window is limited from -5 to 5 in x and y direction}

\item{cexSteps}{size of the stepsTexts}

\item{cexProcent}{size of the procent texts}

\item{parValues}{par(..) values to be called for each chart}

\item{cexPoints}{size of the points}
}
\description{
Spider chart function
}
